chore: clean up tracked pyc files, add analysis scripts and docs
- Remove all tracked __pycache__/*.pyc files (now properly gitignored)
- Add README, analyze_long/short_patterns.py, build cache scripts
- Add docs/superpowers plans and specs
- Update bot source files for latest IBKR migration work

+++ bot/intraday/types.py
@@ -35,7 +35,7 @@ class Signal:
 class Position:
     ticker: str
     direction: Direction
-    shares: int
+    shares: float
     entry_price: float
     stop_price: float
     target_price: float
@@ -61,7 +61,7 @@ class TradeRecord:
     direction: Direction
     entry_time: datetime
     entry_price: float
-    shares: int
+    shares: float
     stop_price: float
     target_price: float
     signals: List[str]
